replacing old importer by new imorter
