+++ myokit/formats/sbml/__init__.py
@@ -7,15 +7,12 @@
 from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 
-from ._importer import SBMLError
-from ._importer import SBMLImporter as OldImporter  # noqa
-from. _new_importer import SBMLImporter as NewImporter
+from ._importer import SBMLImporter, SBMLError  # noqa
 
 
 # Importers
 _importers = {
-    'sbml': NewImporter,
-    'sbmlOld': OldImporter,
+    'sbml': SBMLImporter,
 }
